requirements.txt
certifi==2019.6.16
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
Django==2.2.4
django-ckeditor==5.7.1
django-cors-headers==3.1.1
django-filter==2.2.0
django-guardian==2.0.0
django-js-asset==1.2.2
django-rest-knox==4.1.0
django-taggit==1.1.0
djangorestframework==3.10.2
factory-boy==2.12.0
Faker==2.0.1
gunicorn==20.0.0
idna==2.8
Markdown==3.1.1
Pillow==6.1.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
